Natural Neighbor
Currently unused, but refactors VNA and eVNA in a scikit-
learn style object. This has been tested to reproduce the
DataFusion tool results, but takes just 4 minutes on a PC
instead of 20. Further, it has been enhanced to allow for
multiprocessing to utilize multiple cores. This has the
potential to speed up proportional to system resources.

It has been tested both on Windows and Linux.